Melhorias do projeto
