+++ CMakeLists.txt
@@ -3,5 +3,5 @@ project(Sistemas_Operacionais)
 
 set(CMAKE_CXX_STANDARD 11)
 
-set(SOURCE_FILES main.cpp)
+set(SOURCE_FILES main.cpp VirtualHD.h Manager.h)
 add_executable(Sistemas_Operacionais ${SOURCE_FILES})
